Change the way Brave-specific ContentSettingsTypes are registered
It seems (I'm not fully sure yet) we can't mimic the parameters the
ContentSettingsType::PLUGINS was registered before, when we use it
along with the ResourceIdentifier to implement Brave Shields. Doing
so would apparently cause different problems:
  - Shields would be disabled (BLOCKED) by default
  - Brave-specific ContentSettingsTypes won't be registered in the
    content_settings::PrefProfiver is configured with EPHEMERAL (in
    the past there was an exception for CST::PLUGINS, not anymore).
  - Shields would only apply to WebUI pages and DevTools.

Therefore this patch changes the way Brave-specific ContentSettingTypes
are registered to use, instead, a CONTENT_SETTING_ALLOW default value,
AllowlistedSchemes() and ContentSettingsInfo::PERSISTENT storage, in an
attempt to get things working closer as what they used to be.