[WIP] More experimental changes on the migration away of ResourceIdentifier
Still experimenting and figuring my way out through the code base, this CL
removes references to ContentSettingsType and removes the ephemeral provider
Brave used before, along with some bits of BravePrefProviderm, which we might
not be needing anymore.

Note: It's possible this is all wrong, this is just part of an experiment
and should be cleanup/squashed/removed before merging to the cr88 branch.