Fix Brave Shields panel to show blocked ads again
Turns out that I can't register Brave-specific types with a default
value CONTENT_SETTING_ALLOW, otherwise Brave will do things like
allowing ads by default, which is not correct as far as I can see
comparing with what we have in the master branch. Instead, we need
to pass CONTENT_SETTING_BLOCK here, just as the PLUGIN type was being
registered when we had ResourceIdentifiers.

Additionally, it seems that renaming "ads" to "bs_names" to keep
strings consistent in brave_shields_constants.h is also causing the
panel not to show the number of ads blocked, even if they are now
being effectively blocked (checked with GDB). Thus, restored the names
there not to have a Shields-specific prefix except in those two cases
that is needed not to collide with other ContentSettingsTypes' names
i.e. Cookies and Javascript.